+++ include/executor.h
@@ -1,8 +1,6 @@
 #ifndef EXECUTOR_H
 # define EXECUTOR_H
 
-# include "defines.h"
-
 /* `<SUMMARY>`:
  * Main entrance point for the execution routine. Creates a context
  * variable which holds information about the current file descriptors
@@ -10,8 +8,9 @@
  * `<PARAM>`:
  * `ast`: root node of the AST;
  * `<RETURN>`:
- * Returns TRUE if the builtin exit function is being called. */
-t_bool	executor_main(t_cmd *ast);
+ * Returns QUIT if the program should quit because of the ft_exit
+ * builtin or NO_QUIT if not. */
+t_quit	executor_main(t_cmd *ast);
 
 /* `<SUMMARY>`:
  * Function to examine the type of the current node being processed
@@ -25,18 +24,34 @@ void	exec_node(t_cmd *ast, t_context *ctx);
  * Function to execute a node. Furthermore, decides if it is going
  * to be a builtin function or another function.
  * `<PARAM>`:
- * `ast`: current node of the AST;
+ * `exec`: current execution node of the AST;
  * `ctx`: context for handling the correct redirection; */
 void	execute_command(t_execcmd *exec, t_context *ctx);
 
 /* `<SUMMARY>`:
  * Function to execute the implemented builtin functions.
  * `<PARAM>`:
- * `ast`: current node of the AST;
+ * `exec`: current execution node of the AST;
  * `ctx`: context for handling the correct redirection;
  * `<RETURN>`:
  * Returns TRUE if it is a builtin and FALSE otherwise. */
 t_bool	execute_builtin(t_execcmd *exec, t_context *ctx);
+
+/* `<SUMMARY>`:
+ * Function to execute a redirection. dups the correct file descriptors
+ * and also calls the heredoc function if needed.
+ * `<PARAM>`:
+ * `redir`: current redirection node of the AST;
+ * `ctx`: context for handling the correct redirection; */
+void	execute_redir(t_redircmd *redir, t_context *ctx);
+
+/* `<SUMMARY>`:
+ * Function to execute a node. Furthermore, decides if it is going
+ * to be a builtin function or another function.
+ * `<PARAM>`:
+ * `pcmd`: current pipe node of the AST;
+ * `ctx`: context for handling the correct redirection; */
+void	execute_pipe(t_pipecmd *pcmd, t_context *ctx);
 
 /* `<SUMMARY>`:
  * copy_context copies the necessary variables during the
@@ -50,6 +65,19 @@ t_bool	execute_builtin(t_execcmd *exec, t_context *ctx);
  * Returns TRUE if it is a builtin and FALSE otherwise. */
 void	copy_context(t_context *ctx, t_context next_ctx);
 
+/* `<SUMMARY>`:
+ * Function to add the pid created by the `fork()` call
+ * to the array of pids in the current context.
+ * `<PARAM>`:
+ * `pid`: child pid to add to the array of pids;
+ * `ctx`: context for handling the correct redirection; */
 void	add_child_pids(pid_t pid, t_context *ctx);
+
+/* `<SUMMARY>`:
+ * Function to wait for all pids stored in the pid array
+ * in the context and to set the exit code accordingly.
+ * `<PARAM>`:
+ * `ctx`: context for handling the correct redirection; */
+void	child_reaper(t_context *ctx);
 
 #endif
