+++ src/builtins/ft_cd.c
@@ -4,7 +4,6 @@
 #include "builtins.h"
 #include "environment.h"
 
-
 /* `<SUMMARY>`:
  * Changes the directory to the given `dir` variable.
  * `<PARAM>`:
