+++ src/executor/exec_helper.c
@@ -1,3 +1,4 @@
+#include "defines.h"
 #include "executor.h"
 
 void	add_child_pids(pid_t pid, t_context *ctx)
@@ -21,4 +22,40 @@ void	copy_context(t_context *ctx, t_context next_ctx)
 		++i;
 		++j;
 	}
+}
+
+/* `<SUMMARY>`:
+ * Function for setting the exit code accordingly to the childs
+ * exit code. WIFEXITED is used to determine the return code
+ * if the child exited normally, WIFSIGNALED determines the
+ * return code if the child terminated because of a signal.
+ * `<PARAM>`:
+ * `status`: return value received by waitpid;
+ * `ctx`: context for handling the correct redirection;
+ * `<RETURN>`:
+ * Nothing. */
+static void	set_child_exit_status(int status, t_context *ctx)
+{
+	if (WIFEXITED(status))
+		ctx->exit_code = WEXITSTATUS(status);
+	else if (WIFSIGNALED(status))
+		ctx->exit_code = WTERMSIG(status) + 128;
+}
+
+void	child_reaper(t_context *ctx)
+{
+	int		i;
+	int		status;
+
+	i = -1;
+	status = 0;
+	while (ctx->pids[++i] != 0)
+	{
+		waitpid(ctx->pids[i], &status, 0);
+	}
+	if (ctx->error == TRUE || ctx->exit_code)
+	{
+		return ;
+	}
+	set_child_exit_status(status, ctx);
 }
