+++ src/executor/exec_heredoc.c
@@ -1,0 +1,8 @@
+#include "defines.h"
+#include "executor.h"
+
+void	execute_heredoc(t_redircmd *redir, t_context *ctx)
+{
+
+}
+
