+++ src/parser/parser_main.c
@@ -1,4 +1,5 @@
 #include "libft.h"
+#include "defines.h"
 #include "parser.h"
 #include "lexer.h"
 #include "expander.h"
