+++ src/signals/signal_handler.c
@@ -1,3 +1,4 @@
+#include "defines.h"
 #include "signals.h"
 #include "utils.h"
 
@@ -23,11 +24,6 @@ static void	cleanup_control_character(t_termios *xterm)
 	}
 	local_termios.c_lflag &= ~(ECHOCTL);
 	tcsetattr(STDOUT_FILENO, TCSAFLUSH, &local_termios);
-}
-
-static void	handle_childs()
-{
-	while (waitpid(-1, NULL, WNOHANG) > 0);
 }
 
 /* `<SUMMARY>`:
@@ -61,7 +57,6 @@ int	user_signal_listener(t_termios *xterm)
 {
 	cleanup_control_character(xterm);
 	universal_handler(SIGINT, termination_handler);
-	universal_handler(SIGCHLD, handle_childs);
 	universal_handler(SIGQUIT, SIG_IGN);
 	return (EXIT_SUCCESS);
 }
