+++ src/utils/double_list.c
@@ -1,3 +1,4 @@
+#include "defines.h"
 #include "lexer.h"
 
 t_dlist	*ft_dlstlast(t_dlist *lst)
